requirements.txt
alembic==1.8.1
anyio==3.6.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
fastapi==0.85.1
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
petl==1.7.11
psycopg2==2.9.4
pydantic==1.10.2
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
zipp==3.9.0
